requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
openpyxl==3.1.2
python-decouple==3.8
django-cors-headers==4.3.1
tqdm==4.67.1
requests==2.32.4
